Fix main navigation overlap with custom logo
This commit adds an overlay of the BeforeNav element from
5.0.3, patched with rt commit 512192c.

The right position of topactions is statically set to allow for
the size of the default RT logo.  When a custom logo is uploaded
not matching the size of the default RT logo, topactions is no
longer correctly positioned.

This not only affects the layout, but the menu overflow, since
rebalanceOverflowMenu calculates based on logo width and not the
position of topactions.  This can cause topactions to overlap the
right menu items.

This commit sets the position of topactions based on the logo size
to fix the layout and menu overflow if a custom logo is uploaded.